requirements.txt.txt
python-telegram-bot==21.0
flask
openpyxl